refactor: remove Whisper dependency from ShadowingScreen
- Changed ShadowingScreen to use DemoSessionViewModel instead of SessionViewModel
- Moved ShadowingData class from SessionViewModel to DemoSessionViewModel
- Removed shadowingData flow from SessionViewModel
- Removed DEMO_MODE toggle from App.kt
- ShadowingScreen now works only with demo/dummy data

This change decouples ShadowingScreen from Whisper.cpp integration,
allowing it to function independently with simulated data.
